Added functionality for clues to point to NPCs
- Added %NPC tag to strings in the JSON files. Using this tag will cause
it to be replaced by an NPC's name when NPC.getSpeech() is called.
- When questioning an NPC, normal responses:
	- Will point the player to a random NPC (not the victim, killer or this
NPC) if the NPC being questioned is the killer, or the clue is a red
herring. 
	- Will point the player towards the killer otherwise.
- When questioning an NPC, improved responses:
	- Will point the player to a random NPC (not the victim, killer or this
NPC) ONLY if the NPC being questioned is the killer (red herring status
is ignored).
	- Will point the player towards the killer otherwise, as well as tell
them which room the killer is in.
- The player's personality now changes when they ask questions.
- Changed console printout of rooms from their ID to their name.
- Added some improved dialogue to Colin.JSON