Fixed status bar colour issue
Fixed one of the 2 colour issues that I mentioned show up in the simple
skin brought in, small changes to the json file fixed it. Spent an hour
trying to bring in the pretty sliders and checkboxes from the other skin
but its significantly harder than it looks since it involves playing
with both the json and the atlas file. Ideally we could just use 2
skins, one for buttons, labels, etc and another for the sliders and
checkboxes.